[RFC] Add auto update functionality on OSX
This commit is a quick hack that shows how auto update could work
on quassel. On OSX and Windows, there is not a common repository
solution like *nix, so people rely on the application to handle
this.

The solution is a slightly modified/CMake'd port of:
http://el-tramo.be/blog/mixing-cocoa-and-qt/

If the idea is accepted, I would like to support Windows as well,
with an approach similar to:
https://github.com/Roxee/qt-roxeemegaup
Note that the mentioned project needs some updates, and
generalization before using within quassel.